requirements.txt
# Core web stack
fastapi==0.115.0
uvicorn[standard]==0.30.6
gradio==4.44.1
httpx==0.27.2
pydantic==2.9.2
pydantic-core==2.23.4
python-multipart==0.0.12
jinja2==3.1.4

# Data & DB
SQLAlchemy==2.0.34
alembic==1.13.2
aiosqlite==0.20.0

# ML & CV
ultralytics==8.3.39
opencv-python==4.10.0.84
numpy==1.26.4
Pillow==10.4.0
scikit-image==0.24.0
torch>=2.1.0,<=2.4.0
 torchvision>=0.16.0,<=0.19.0
onnxruntime==1.19.2
pytesseract==0.3.13

# Utils
python-dotenv==1.0.1
loguru==0.7.2
cachetools==5.5.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx[cli]==0.27.2
